requirements.txt
streamlit==1.28.1
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.1
plotly==5.17.0
onnxruntime==1.16.1
scipy==1.11.3
